requirements.txt
# HERMES AI Voice Agent System - Consolidated Python Dependencies
# Review Date: 2025-09-14
# This version deduplicates previous requirements files and aligns with CI/dev tooling.

# --- Core API & Async Stack ---
fastapi==0.115.5
uvicorn[standard]==0.32.1
pydantic==2.10.3
pydantic-settings==2.6.1
python-multipart==0.0.18
httpx==0.28.1
websockets==14.1
mcp>=1.0.0

# --- Data & Caching Layer ---
sqlalchemy[asyncio]==2.0.36
asyncpg==0.30.0
redis[hiredis]==5.1.2

# --- AI, Voice, and Core Utilities ---
openai==1.3.7
openai-whisper==20231117
torch==2.1.1
torchaudio==2.1.1
numpy==1.26.4
requests==2.32.5

# --- Integrations & External Services ---
requests-oauthlib==1.3.1
supabase==2.0.2
postgrest==0.13.0
PyGithub==1.59.1
pygit2==1.13.3
mem0ai==0.0.7
stripe==11.2.0
zapier-platform-cli==15.0.0
playwright==1.40.0

# --- Security, Auth, and Compliance ---
cryptography==44.0.0
python-jose[cryptography]==3.3.0
pyjwt[crypto]==2.10.1
passlib[bcrypt]==1.7.4
python-dotenv==1.0.1
prometheus-client==0.19.0
structlog==23.2.0
python-json-logger==2.0.7

# --- Observability & Deployment ---
gunicorn==21.2.0
docker==6.1.3
google-cloud-run==0.10.0
google-cloud-logging==3.8.0

# --- Document Generation ---
reportlab==4.0.7
python-docx==1.1.0
pdfplumber==0.9.0

# --- Quality & Testing Tooling ---
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
coverage[toml]==7.4.4
black==23.11.0
flake8==6.1.0
mypy==1.7.1
isort==5.12.0

